Added support for modular b2.
This adds explicit specification of all library dependencies, which is
necessary for modular b2 builds and also for tracking usage requirements
from the dependencies.

This also introduces a new boost_sync_with_support target, which
provides a dependency on the main Boost.Sync library, as well as any
additional dependencies that are needed for support headers.